fix remove travel address

+++ app/Http/Controllers/Travels/TravelsController.php
@@ -639,10 +639,12 @@ class TravelsController extends Controller
         }
 
         $ids = Arr::pluck($travelAddr, 'id');
-        foreach ($travel->travelAddress as $oldAddr) {
-            if (!in_array($oldAddr->id, $ids)) {
-                $currentTravelAddr = $travel->travelAddress()->findOrFail($oldAddr->id);
-                $currentTravelAddr->delete();
+        if($travel->travelAddress()->exists()) {
+            foreach ($travel->travelAddress as $oldAddr) {
+                if (!in_array($oldAddr->id, $ids)) {
+                    $currentTravelAddr = $travel->travelAddress()->findOrFail($oldAddr->id);
+                    $currentTravelAddr->delete();
+                }
             }
         }
